refactor: déplace tous les hooks dans le dossier hooks
