+++ site/source/components/BarChart.tsx
@@ -4,10 +4,10 @@ import { useTranslation } from 'react-i18next'
 import { animated, config, useSpring } from 'react-spring'
 import { styled, useTheme } from 'styled-components'
 
-import useDisplayOnIntersecting from '@/components/utils/useDisplayOnIntersecting'
 import { Emoji } from '@/design-system/emoji'
 import { Spacing } from '@/design-system/layout'
 import { Body, SmallBody } from '@/design-system/typography/paragraphs'
+import useDisplayOnIntersecting from '@/hooks/useDisplayOnIntersecting'
 
 import { DisableAnimationContext } from './utils/DisableAnimationContext'
