+++ site/source/components/SimulateurOrAssistantPage.tsx
@@ -4,7 +4,6 @@ import { useLocation } from 'react-router-dom'
 import { styled } from 'styled-components'
 
 import Meta from '@/components/utils/Meta'
-import useSearchParamsSimulationSharing from '@/components/utils/useSearchParamsSimulationSharing'
 import { Chip } from '@/design-system'
 import { Emoji } from '@/design-system/emoji'
 import { H1 } from '@/design-system/typography/heading'
@@ -14,6 +13,7 @@ import {
 	useCurrentSimulatorData,
 } from '@/hooks/useCurrentSimulatorData'
 import { useIsEmbedded } from '@/hooks/useIsEmbedded'
+import useSearchParamsSimulationSharing from '@/hooks/useSearchParamsSimulationSharing'
 import useSimulationConfig from '@/hooks/useSimulationConfig'
 import { Simulation } from '@/store/reducers/simulation.reducer'
 import { situationSelector } from '@/store/selectors/simulationSelectors'
