+++ site/source/pages/assistants/declaration-charges-sociales-independant/components/ImpositionSection.tsx
@@ -9,9 +9,9 @@ import RuleInput from '@/components/conversation/RuleInput'
 import { Condition } from '@/components/EngineValue/Condition'
 import { WhenAlreadyDefined } from '@/components/EngineValue/WhenAlreadyDefined'
 import { FromTop } from '@/components/ui/animate'
-import useYear from '@/components/utils/useYear'
 import { H2 } from '@/design-system/typography/heading'
 import { SmallBody } from '@/design-system/typography/paragraphs'
+import useYear from '@/hooks/useYear'
 import { enregistreLaRéponse } from '@/store/actions/actions'
 
 import { SimpleField } from '../../components/Fields'
