+++ site/source/pages/simulateurs/cessation-activité/CessationActivité.tsx
@@ -3,8 +3,8 @@ import { useSelector } from 'react-redux'
 
 import SimulateurWarning from '@/components/SimulateurWarning'
 import Simulation from '@/components/Simulation'
-import { useParamsFromSituation } from '@/components/utils/useSearchParamsSimulationSharing'
 import { Body } from '@/design-system/typography/paragraphs'
+import { useParamsFromSituation } from '@/hooks/useSearchParamsSimulationSharing'
 import useSimulatorsData from '@/hooks/useSimulatorsData'
 import { useSiteUrl } from '@/hooks/useSiteUrl'
 import { CessationActivitéGoals } from '@/pages/simulateurs/cessation-activité/Goals'
